feat: allow Ns in the sample barcode
This is also useful if the barcodes are of different length, so we can
pad the shorter barcodes with Ns
